Fix 'Image Not Found' Error When Including Framework via SPM (#13)
- When we include a framework using SPM (Swift Package Manager), it
doesn't have a fixed bundle identifier that allows us to directly access
its resources. Therefore, we need to use the Bundle.module property to
access resources within the framework when it's included via SPM.
- However, if we include a framework directly in the app (as
demonstrated in the example app within this project), the Bundle.module
approach won't work. In such cases, we need to obtain the bundle using
its identifier in order to access the resources.