fix: Player orientation not changed on fullscreen below iOS 16 (#34)
- Previously, player orientation did not switch to landscape on
fullscreen for iOS versions below 16. Below iOS 16, we set the
orientation using UIDevice but incorrectly used
`UIInterfaceOrientationMask` instead of the correct
`UIInterfaceOrientation` rawValue. As a result, the orientation did not
change.
- This commit resolves the issue by introducing a method via an
extension to obtain the equivalent UIInterfaceOrientation from
UIInterfaceOrientationMask then used to retrieve the correct orientation
and set its raw value on UIDevice.