fix:  Upgrade Sentry Cocoa dependency to 8.50.0
The previous Sentry Cocoa version (8.36.0) caused compilation errors when building with Xcode 16.3. This issue did not occur when the SDK was integrated via Swift Package Manager or as an Xcode framework, because those integrations used .upToNextMajor versioning which allowed resolving to a newer, compatible version. To fix the issue when integrating via CocoaPods, the Podspec dependency has been updated to Sentry version 8.50.0, which includes necessary fixes for compatibility with Xcode 16.3.